+++ src/router.js
@@ -1,4 +1,4 @@
-import Model from "./model.js";
+import Controller from "./controller.js";
 import tech from "./views/tech.js";
 import article from "./views/article.js";
 import design from "./views/design.js";
@@ -12,27 +12,6 @@ const routes = [
 
 const pathToRegex = (path) =>
   new RegExp("^" + path.replace(/\//g, "\\/").replace(/:\w+/g, "(.+)") + "$");
-
-const getArticles = (matchPath) => {
-  const view = new matchPath.route.view();
-  const id = matchPath.result[1];
-  const model = new Model();
-
-  model.then((data) => {
-    const dataType = !id
-      ? matchPath.route.view.name.toLowerCase()
-      : data.find((item) => item.id === id).type;
-
-    const dataGroupByType = data.filter((item) => item.type === dataType);
-
-    if (!id) {
-      view.getData(dataGroupByType);
-    } else {
-      const article = data.find((item) => item.id === id);
-      view.getData(article);
-    }
-  });
-};
 
 const router = () => {
   const { pathname } = window.location;
@@ -51,7 +30,8 @@ const router = () => {
     contents.innerHTML = page404;
     return;
   } else {
-    getArticles(match);
+    const controller = new Controller(match.route.view);
+    controller.getArticles(match);
   }
 };
